app/requirements.txt
streamlit
opencv-python-headless
Pillow
onnxruntime
torch
transformers